added the first test
