+++ Amazon.sln.DotSettings.user
@@ -1,6 +1,6 @@
 ﻿<wpf:ResourceDictionary xml:space="preserve" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:s="clr-namespace:System;assembly=mscorlib" xmlns:ss="urn:shemas-jetbrains-com:settings-storage-xaml" xmlns:wpf="http://schemas.microsoft.com/winfx/2006/xaml/presentation">
 	<s:String x:Key="/Default/Environment/UnitTesting/UnitTestSessionStore/Sessions/=3c4f20e6_002Dd31b_002D4088_002Db349_002D0e42e0578138/@EntryIndexedValue">&lt;SessionState ContinuousTestingMode="0" IsActive="True" Name="Setup" xmlns="urn:schemas-jetbrains-com:jetbrains-ut-session"&gt;&#xD;
   &lt;TestAncestor&gt;&#xD;
-    &lt;TestId&gt;NUnit3x::AF2E10BF-B7B7-4169-9D0F-F4AF3D1CDD5C::.NETFramework,Version=v4.8::Amazon.Tests.SomeTest.Setup&lt;/TestId&gt;&#xD;
+    &lt;TestId&gt;NUnit3x::AF2E10BF-B7B7-4169-9D0F-F4AF3D1CDD5C::.NETFramework,Version=v4.8::Amazon.Tests.ImageComparisonTests.SearchForProduct&lt;/TestId&gt;&#xD;
   &lt;/TestAncestor&gt;&#xD;
 &lt;/SessionState&gt;</s:String></wpf:ResourceDictionary>
